apps/bot/pyproject.toml
[project]
name = "sharpedge-bot"
version = "0.1.0"
description = "SharpEdge Discord bot - Sports betting intelligence"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.0",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "stripe>=11.0",
    "httpx>=0.27",
    "openai-agents>=0.0.7",
    "apscheduler>=3.10",
    "sharpedge-shared",
    "sharpedge-db",
    "sharpedge-odds",
]

[project.scripts]
sharpedge-bot = "sharpedge_bot.main:main"

[tool.uv.sources]
sharpedge-shared = { workspace = true }
sharpedge-db = { workspace = true }
sharpedge-odds = { workspace = true }
